refactor: unify test/util/wallet.h with wallet/test/util.h
files share the same purpose, and we should not have wallet code inside the test directory.

This later is needed to use wallet util functions in the bench and test binaries without being forced to duplicate them.

(cherry picked from commit ee7a984f85015b610be4929b7c35cb501c1fbf7c)
