+++ src/Makefile.test_util.include
@@ -24,9 +24,13 @@ TEST_UTIL_H = \
   test/util/str.h \
   test/util/transaction_utils.h \
   test/util/txmempool.h \
-  test/util/wallet.h \
   test/util/validation.h \
   test/util/xoroshiro128plusplus.h
+
+if ENABLE_WALLET
+TEST_UTIL_H += \
+  wallet/test/util.h
+endif # ENABLE_WALLET
 
 libtest_util_a_CPPFLAGS = $(AM_CPPFLAGS) $(BITCOIN_INCLUDES) $(BOOST_CPPFLAGS)
 libtest_util_a_CXXFLAGS = $(AM_CXXFLAGS) $(PIE_FLAGS)
@@ -44,5 +48,9 @@ libtest_util_a_SOURCES = \
   test/util/transaction_utils.cpp \
   test/util/txmempool.cpp \
   test/util/validation.cpp \
-  test/util/wallet.cpp \
   $(TEST_UTIL_H)
+
+if ENABLE_WALLET
+libtest_util_a_SOURCES += \
+  wallet/test/util.cpp
+endif # ENABLE_WALLET
