+++ src/bench/wallet_create_tx.cpp
@@ -8,14 +8,15 @@
 #include <consensus/merkle.h>
 #include <node/context.h>
 #include <test/util/setup_common.h>
-#include <test/util/wallet.h>
 #include <validation.h>
 #include <wallet/spend.h>
+#include <wallet/test/util.h>
 #include <wallet/wallet.h>
 
 using wallet::CWallet;
 using wallet::CreateMockWalletDatabase;
 using wallet::DBErrors;
+using wallet::getNewDestination;
 using wallet::WALLET_FLAG_DESCRIPTORS;
 
 struct TipBlock
